+++ src/Bit/CountLeadingZeros.h
@@ -7,11 +7,11 @@
 #include "BitLength.h"
 #include "../Concepts/Integral.h"
 #include "../Traits/IntegralTrait.h"
-
+#include "../ArithmeticType.h"
 
 
 template <concepts::Integral T>
-[[nodiscard]] constexpr int CountLeadingZeros(T x) noexcept {
+[[nodiscard]] constexpr usize CountLeadingZeros(T x) noexcept {
     return traits::IntegralTrait<T>::NumBits - ::BitLength(x);
 }
